# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MigrateRequest {
  /** The source code to migrate */
  code: string;
  /** Optional filename hint (e.g. index.ts) */
  filename?: string;
}

export interface TransformDetail {
  /** Category: imports | connection | keypair | publickey | transaction | buffer | lamports */
  category: string;
  original: string;
  transformed: string;
  flaggedForAI: boolean;
  line?: number;
}

export type MigrationStatsByCategory = { [key: string]: number };

export interface MigrationStats {
  totalChanges: number;
  automaticChanges: number;
  aiRequiredChanges: number;
  coveragePercent: number;
  byCategory: MigrationStatsByCategory;
}

export interface MigrateResult {
  transformedCode: string;
  transforms: TransformDetail[];
  stats: MigrationStats;
  sessionId: string;
}

export interface CoverageCategory {
  name: string;
  description: string;
  coveragePercent: number;
  automated: boolean;
  transformCount: number;
}

export interface CoverageStats {
  overall: number;
  categories: CoverageCategory[];
  totalTransforms: number;
  reposTestedCount: number;
  reposTested: string[];
}

export interface MigrationExample {
  id: string;
  title: string;
  category: string;
  before: string;
  after: string;
  description: string;
  flaggedForAI: boolean;
}

export interface ExampleList {
  examples: MigrationExample[];
}

export interface Session {
  id: string;
  createdAt: string;
  filename?: string;
  originalCode: string;
  transformedCode: string;
  stats: MigrationStats;
}

export interface SessionList {
  sessions: Session[];
}