# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Solana Kit Migrator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: migrate
    description: Code migration operations
  - name: stats
    description: Migration statistics and coverage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /migrate:
    post:
      operationId: migrateCode
      tags: [migrate]
      summary: Migrate Solana web3.js v1 code to @solana/kit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrateRequest"
      responses:
        "200":
          description: Migration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrateResult"

  /coverage:
    get:
      operationId: getCoverage
      tags: [stats]
      summary: Get coverage breakdown by migration category
      responses:
        "200":
          description: Coverage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverageStats"

  /examples:
    get:
      operationId: getExamples
      tags: [stats]
      summary: Get before/after migration examples
      responses:
        "200":
          description: Migration examples
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExampleList"

  /sessions:
    get:
      operationId: listSessions
      tags: [migrate]
      summary: List recent migration sessions
      responses:
        "200":
          description: List of recent migration sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [migrate]
      summary: Get a specific migration session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MigrateRequest:
      type: object
      properties:
        code:
          type: string
          description: The source code to migrate
        filename:
          type: string
          description: Optional filename hint (e.g. index.ts)
      required:
        - code

    TransformDetail:
      type: object
      properties:
        category:
          type: string
          description: "Category: imports | connection | keypair | publickey | transaction | buffer | lamports"
        original:
          type: string
        transformed:
          type: string
        flaggedForAI:
          type: boolean
        line:
          type: integer
      required:
        - category
        - original
        - transformed
        - flaggedForAI

    MigrateResult:
      type: object
      properties:
        transformedCode:
          type: string
        transforms:
          type: array
          items:
            $ref: "#/components/schemas/TransformDetail"
        stats:
          $ref: "#/components/schemas/MigrationStats"
        sessionId:
          type: string
      required:
        - transformedCode
        - transforms
        - stats
        - sessionId

    MigrationStats:
      type: object
      properties:
        totalChanges:
          type: integer
        automaticChanges:
          type: integer
        aiRequiredChanges:
          type: integer
        coveragePercent:
          type: number
        byCategory:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalChanges
        - automaticChanges
        - aiRequiredChanges
        - coveragePercent
        - byCategory

    CoverageCategory:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coveragePercent:
          type: integer
        automated:
          type: boolean
        transformCount:
          type: integer
      required:
        - name
        - description
        - coveragePercent
        - automated
        - transformCount

    CoverageStats:
      type: object
      properties:
        overall:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CoverageCategory"
        totalTransforms:
          type: integer
        reposTestedCount:
          type: integer
        reposTested:
          type: array
          items:
            type: string
      required:
        - overall
        - categories
        - totalTransforms
        - reposTestedCount
        - reposTested

    MigrationExample:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        before:
          type: string
        after:
          type: string
        description:
          type: string
        flaggedForAI:
          type: boolean
      required:
        - id
        - title
        - category
        - before
        - after
        - description
        - flaggedForAI

    ExampleList:
      type: object
      properties:
        examples:
          type: array
          items:
            $ref: "#/components/schemas/MigrationExample"
      required:
        - examples

    Session:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        filename:
          type: string
        originalCode:
          type: string
        transformedCode:
          type: string
        stats:
          $ref: "#/components/schemas/MigrationStats"
      required:
        - id
        - createdAt
        - transformedCode
        - originalCode
        - stats

    SessionList:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
      required:
        - sessions