# lib/api-zod/src/generated/types/coverageCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */

export interface CoverageCategory {
  name: string;
  description: string;
  coveragePercent: number;
  automated: boolean;
  transformCount: number;
}