# lib/api-zod/src/generated/types/coverageStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageCategory } from "./coverageCategory";

export interface CoverageStats {
  overall: number;
  categories: CoverageCategory[];
  totalTransforms: number;
  reposTestedCount: number;
  reposTested: string[];
}