# lib/api-zod/src/generated/types/exampleList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
import type { MigrationExample } from "./migrationExample";

export interface ExampleList {
  examples: MigrationExample[];
}