# lib/api-zod/src/generated/types/migrateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */

export interface MigrateRequest {
  /** The source code to migrate */
  code: string;
  /** Optional filename hint (e.g. index.ts) */
  filename?: string;
}