# lib/api-zod/src/generated/types/migrateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
import type { MigrationStats } from "./migrationStats";
import type { TransformDetail } from "./transformDetail";

export interface MigrateResult {
  transformedCode: string;
  transforms: TransformDetail[];
  stats: MigrationStats;
  sessionId: string;
}