# lib/api-zod/src/generated/types/migrationExample.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */

export interface MigrationExample {
  id: string;
  title: string;
  category: string;
  before: string;
  after: string;
  description: string;
  flaggedForAI: boolean;
}