# lib/api-zod/src/generated/types/migrationStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
import type { MigrationStatsByCategory } from "./migrationStatsByCategory";

export interface MigrationStats {
  totalChanges: number;
  automaticChanges: number;
  aiRequiredChanges: number;
  coveragePercent: number;
  byCategory: MigrationStatsByCategory;
}