# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */
import type { MigrationStats } from "./migrationStats";

export interface Session {
  id: string;
  createdAt: Date;
  filename?: string;
  originalCode: string;
  transformedCode: string;
  stats: MigrationStats;
}