# lib/api-zod/src/generated/types/transformDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solana Kit Migrator API
 * OpenAPI spec version: 0.1.0
 */

export interface TransformDetail {
  /** Category: imports | connection | keypair | publickey | transaction | buffer | lamports */
  category: string;
  original: string;
  transformed: string;
  flaggedForAI: boolean;
  line?: number;
}